Fix or silence (msvc) compiler warnings about unused formal parameters, unused local variables, narrowing conversions, constant conditional expressions, user-defined comma operators, name hiding, ambiguous overloads, and signed-unsigned-mismatches
(cherry picked from commit 8bf4a81a3bbff4404bb3a678b80d1569078091ac)

+++ test/all_planar_input_files_test.cpp
@@ -149,7 +149,7 @@ int test_graph(const std::string& dimacs_filename)
 
   // This edge_updater will automatically update the interior edge
   // index of the graph as edges are created.
-  edge_visitor_t edge_updater(get(edge_index, g), num_edges(g));
+  edge_visitor_t edge_updater(get(edge_index, g), static_cast<int>(num_edges(g)));
 
   // The input graph may not be maximal planar, but the Chrobak-Payne straight
   // line drawing needs a maximal planar graph as input. So, we make a copy of
